+++ packages/hyperion-autologging-visualizer/src/index.ts
@@ -3,18 +3,5 @@
  */
 
 'use strict';
-import * as Types from "@hyperion/hyperion-util/src/Types";
-import * as AutoLogging from "@hyperion/hyperion-autologging/src/AutoLogging";
-import { Channel } from "@hyperion/hyperion-channel/src/Channel";
-import * as ElementTextTooltip from "./component/ElementTextTooltip.react";
-
-export type InitOptions = Types.Options<
-  Pick<AutoLogging.InitOptions, 'flowletManager'> &
-  {
-    channel: Channel<AutoLogging.ALChannelEvent>
-  }
->;
-
-export function init(options: InitOptions) {
-  ElementTextTooltip.init(options);
-}
+export { ALGraph, AL_GRAPH_SCRATCH_NAMESPACE, ALGraphDefaultDynamicOptions } from "./component/ALGraph";
+export { ALGraphInfo } from "./component/ALGraphInfo.react";
