+++ packages/hyperion-channel/src/index.ts
@@ -2,5 +2,5 @@
  * Copyright (c) Meta Platforms, Inc. and affiliates. All Rights Reserved.
  */
 
-export * from "./assert";
-// export { setAssertLoggerOptions } from "./assert";
+'use strict';
+export { PipeableEmitter, Channel, PausableChannel } from "./Channel";
