+++ packages/hyperion-devtools/src/index.ts
@@ -1,6 +1,3 @@
 /**
  * Copyright (c) Meta Platforms, Inc. and affiliates. All Rights Reserved.
  */
-
-export * from "./assert";
-// export { setAssertLoggerOptions } from "./assert";
