+++ packages/hyperion-dom/src/index.ts
@@ -1,0 +1,15 @@
+/**
+ * Copyright (c) Meta Platforms, Inc. and affiliates. All Rights Reserved.
+ */
+
+'use strict';
+export * as IEvent from "./IEvent";
+export * as IEventTarget from "./IEventTarget";
+// export * as INode from "./INode";
+export * as IElement from "./IElement";
+export * as IHTMLElement from "./IHTMLElement";
+export * as IHTMLInputElement from "./IHTMLInputElement";
+export * as ICSSStyleDeclaration from "./ICSSStyleDeclaration";
+// export * as IGlobalEventHandlers from "./IGlobalEventHandlers";
+export * as IWindow from "./IWindow";
+// export * as IXMLHttpRequest from "./IXMLHttpRequest";
