+++ packages/hyperion-react/src/index.ts
@@ -1,0 +1,8 @@
+/**
+ * Copyright (c) Meta Platforms, Inc. and affiliates. All Rights Reserved.
+ */
+
+'use strict';
+export * as IReact from "./IReact";
+export * as IReactDOM from "./IReactDOM";
+export * as IReactComponent from "./IReactComponent"
