+++ packages/hyperion-test-and-set/src/index.ts
@@ -2,5 +2,5 @@
  * Copyright (c) Meta Platforms, Inc. and affiliates. All Rights Reserved.
  */
 
-export * from "./assert";
-// export { setAssertLoggerOptions } from "./assert";
+'use strict';
+export { default as TestAndSet } from "./TestAndSet";
